+++ src/lib/output/themes/default/partials/toolbar.tsx
@@ -2,7 +2,6 @@ import type { Reflection } from "../../../../models";
 import { JSX } from "../../../../utils";
 import type { PageEvent } from "../../../events";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
-import { icons } from "./icon";
 
 export const toolbar = (context: DefaultThemeRenderContext, props: PageEvent<Reflection>) => (
     <header class="tsd-page-toolbar">
@@ -10,7 +9,7 @@ export const toolbar = (context: DefaultThemeRenderContext, props: PageEvent<Ref
             <div class="table-cell" id="tsd-search" data-base={context.relativeURL("./")}>
                 <div class="field">
                     <label for="tsd-search-field" class="tsd-widget search no-caption">
-                        {icons.search()}
+                        {context.icons.search()}
                     </label>
                     <input type="text" id="tsd-search-field" aria-label="Search" />
                 </div>
@@ -27,7 +26,7 @@ export const toolbar = (context: DefaultThemeRenderContext, props: PageEvent<Ref
 
             <div class="table-cell" id="tsd-widgets">
                 <a href="#" class="tsd-widget menu no-caption" data-toggle="menu" aria-label="Menu">
-                    {icons.menu()}
+                    {context.icons.menu()}
                 </a>
             </div>
         </div>