+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 -   The `DEBUG_SEARCH_WEIGHTS` global variable can now be set on `window` to add search scoring information in the search results.
+-   TypeDoc's icons are now available on `DefaultThemeRenderContext.icons` for use/modification by themes.
 
 ## v0.23.4 (2022-07-02)